requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.8
rq==1.16.2
boto3==1.35.23
Pillow==10.4.0
ImageHash==4.3.1
loguru==0.7.2
rapidfuzz==3.9.6
python-dotenv==1.0.1
